requirements.txt
flask
python-dotenv
airtable-python-wrapper
apscheduler
jinja2
requests
google-api-python-client